#329 move account selection to own component
